+++ crates/rust-client/Cargo.toml
@@ -60,14 +60,15 @@ toml = { version = "0.8", optional = true }
 ignored = ["getrandom"]
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
+async-lock = { version = "3.4.0", default-features = false } # TODO(Maks) for POC only, to be removed
 js-sys = "0.3"
+pollster = { version = "0.4.0", default-features = false }
 sqlite-wasm-rs = { version = "0.3", default-features = false, features = ["bundled"], optional = true}
 web-sys = { version = "0.3", features = ["console", "Window", "Storage", "FileSystemDirectoryHandle",
     "FileSystemFileHandle",
     "FileSystemWritableFileStream",
     "FileSystemGetDirectoryOptions",
     "FileSystemGetFileOptions",
-    "FileSystemCreateWritableOptions",
     "Blob",
     "Navigator",
     "StorageManager",
