+++ crates/rust-client/src/store/sqlite_store/mod.rs
@@ -12,17 +12,41 @@ mod note;
 mod sync;
 mod transaction;
 
+use alloc::{
+    boxed::Box,
+    collections::{BTreeMap, BTreeSet},
+    string::ToString,
+    vec::Vec,
+};
+
 #[cfg(not(target_arch = "wasm32"))]
-mod store_native;
-#[cfg(not(target_arch = "wasm32"))]
-use db_management::SqlitePool;
+use std::path::PathBuf;
 
 #[cfg(target_arch = "wasm32")]
-mod store_web;
-#[cfg(target_arch = "wasm32")]
-pub use store_web::WebStore;
+type PathBuf = &'static str;
 
-use rusqlite::types::Value;
+use crate::store::sqlite_store::db_management::{SqlitePool, utils::apply_migrations};
+use miden_objects::{
+    Digest, Word,
+    account::{Account, AccountCode, AccountHeader, AccountId},
+    block::{BlockHeader, BlockNumber},
+    crypto::merkle::{InOrderIndex, MmrPeaks},
+    note::{NoteTag, Nullifier},
+};
+use rusqlite::{Connection, types::Value};
+
+#[cfg(any(feature = "tonic", feature = "web-tonic"))]
+use tonic::async_trait;
+
+use crate::store::{
+    AccountRecord, AccountStatus, InputNoteRecord, NoteFilter, OutputNoteRecord,
+    PartialBlockchainFilter, Store, TransactionFilter,
+};
+use crate::{
+    store::StoreError,
+    sync::{NoteTagRecord, StateSyncUpdate},
+    transaction::{TransactionRecord, TransactionStoreUpdate},
+};
 
 // SQLITE STORE
 // ================================================================================================
@@ -32,8 +56,306 @@ use rusqlite::types::Value;
 ///
 /// Current table definitions can be found at `store.sql` migration file.
 pub struct SqliteStore {
-    #[cfg(not(target_arch = "wasm32"))]
     pub(crate) pool: SqlitePool,
+}
+
+impl SqliteStore {
+    // CONSTRUCTORS
+    // --------------------------------------------------------------------------------------------
+
+    /// Returns a new instance of [Store] instantiated with the specified configuration options.
+    pub async fn new(database_filepath: PathBuf) -> Result<Self, StoreError> {
+        #[cfg(not(target_arch = "wasm32"))]
+        let pool = {
+            let pool = SqlitePool::connect(database_filepath.clone())
+                .await
+                .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+            let conn = pool.get().await.map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+            let _ = conn
+                .interact(apply_migrations)
+                .await
+                .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+            pool
+        };
+
+        #[cfg(target_arch = "wasm32")]
+        let pool = {
+            let pool = SqlitePool::connect(database_filepath)
+                .await
+                .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+            pool.interact(apply_migrations)
+                .await
+                .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+            pool
+        };
+
+        Ok(SqliteStore { pool })
+    }
+
+    /// Interacts with the database by executing the provided function on a connection from the
+    /// pool.
+    ///
+    /// This function is a helper method which simplifies the process of making queries to the
+    /// database. It acquires a connection from the pool and executes the provided function,
+    /// returning the result.
+    #[cfg(target_arch = "wasm32")]
+    async fn interact_with_connection<F, R>(&self, f: F) -> Result<R, StoreError>
+    where
+        F: FnOnce(&mut Connection) -> Result<R, StoreError> + Send + 'static,
+        R: Send + 'static,
+    {
+        self.pool
+            .interact(f)
+            .await
+            .map_err(|err| StoreError::DatabaseError(err.to_string()))
+    }
+
+    #[cfg(not(target_arch = "wasm32"))]
+    async fn interact_with_connection<F, R>(&self, f: F) -> Result<R, StoreError>
+    where
+        F: FnOnce(&mut Connection) -> Result<R, StoreError> + Send + 'static,
+        R: Send + 'static,
+    {
+        self.pool
+            .get()
+            .await
+            .map_err(|err| StoreError::DatabaseError(err.to_string()))?
+            .interact(f)
+            .await
+            .map_err(|err| StoreError::DatabaseError(err.to_string()))?
+    }
+}
+
+// SQLite implementation of the Store trait
+//
+// To simplify, all implementations rely on inner SqliteStore functions that map 1:1 by name
+// This way, the actual implementations are grouped by entity types in their own sub-modules
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+impl Store for SqliteStore {
+    fn get_current_timestamp(&self) -> Option<u64> {
+        let now = chrono::Utc::now();
+        Some(u64::try_from(now.timestamp()).expect("timestamp is always after epoch"))
+    }
+
+    async fn get_note_tags(&self) -> Result<Vec<NoteTagRecord>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_note_tags).await
+    }
+
+    async fn get_unique_note_tags(&self) -> Result<BTreeSet<NoteTag>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_unique_note_tags).await
+    }
+
+    async fn add_note_tag(&self, tag: NoteTagRecord) -> Result<bool, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::add_note_tag(conn, tag))
+            .await
+    }
+
+    async fn remove_note_tag(&self, tag: NoteTagRecord) -> Result<usize, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::remove_note_tag(conn, tag))
+            .await
+    }
+
+    async fn get_sync_height(&self) -> Result<BlockNumber, StoreError> {
+        self.interact_with_connection(SqliteStore::get_sync_height).await
+    }
+
+    async fn apply_state_sync(&self, state_sync_update: StateSyncUpdate) -> Result<(), StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::apply_state_sync(conn, state_sync_update)
+        })
+        .await
+    }
+
+    async fn get_transactions(
+        &self,
+        transaction_filter: TransactionFilter,
+    ) -> Result<Vec<TransactionRecord>, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_transactions(conn, &transaction_filter)
+        })
+        .await
+    }
+
+    async fn apply_transaction(&self, tx_update: TransactionStoreUpdate) -> Result<(), StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::apply_transaction(conn, &tx_update))
+            .await
+    }
+
+    async fn get_input_notes(
+        &self,
+        filter: NoteFilter,
+    ) -> Result<Vec<InputNoteRecord>, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::get_input_notes(conn, &filter))
+            .await
+    }
+
+    async fn get_output_notes(
+        &self,
+        note_filter: NoteFilter,
+    ) -> Result<Vec<OutputNoteRecord>, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::get_output_notes(conn, &note_filter))
+            .await
+    }
+
+    async fn upsert_input_notes(&self, notes: &[InputNoteRecord]) -> Result<(), StoreError> {
+        let notes = notes.to_vec();
+        self.interact_with_connection(move |conn| SqliteStore::upsert_input_notes(conn, &notes))
+            .await
+    }
+
+    async fn insert_block_header(
+        &self,
+        block_header: &BlockHeader,
+        partial_blockchain_peaks: MmrPeaks,
+        has_client_notes: bool,
+    ) -> Result<(), StoreError> {
+        let block_header = block_header.clone();
+        self.interact_with_connection(move |conn| {
+            SqliteStore::insert_block_header(
+                conn,
+                &block_header,
+                &partial_blockchain_peaks,
+                has_client_notes,
+            )
+        })
+        .await
+    }
+
+    async fn prune_irrelevant_blocks(&self) -> Result<(), StoreError> {
+        self.interact_with_connection(SqliteStore::prune_irrelevant_blocks).await
+    }
+
+    async fn get_block_headers(
+        &self,
+        block_numbers: &BTreeSet<BlockNumber>,
+    ) -> Result<Vec<(BlockHeader, bool)>, StoreError> {
+        let block_numbers = block_numbers.clone();
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_block_headers(conn, &block_numbers)
+        })
+        .await
+    }
+
+    async fn get_tracked_block_headers(&self) -> Result<Vec<BlockHeader>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_tracked_block_headers).await
+    }
+
+    async fn get_partial_blockchain_nodes(
+        &self,
+        filter: PartialBlockchainFilter,
+    ) -> Result<BTreeMap<InOrderIndex, Digest>, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_partial_blockchain_nodes(conn, &filter)
+        })
+        .await
+    }
+
+    async fn insert_partial_blockchain_nodes(
+        &self,
+        nodes: &[(InOrderIndex, Digest)],
+    ) -> Result<(), StoreError> {
+        let nodes = nodes.to_vec();
+        self.interact_with_connection(move |conn| {
+            SqliteStore::insert_partial_blockchain_nodes(conn, &nodes)
+        })
+        .await
+    }
+
+    async fn get_partial_blockchain_peaks_by_block_num(
+        &self,
+        block_num: BlockNumber,
+    ) -> Result<MmrPeaks, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_partial_blockchain_peaks_by_block_num(conn, block_num)
+        })
+        .await
+    }
+
+    async fn insert_account(
+        &self,
+        account: &Account,
+        account_seed: Option<Word>,
+    ) -> Result<(), StoreError> {
+        let account = account.clone();
+
+        self.interact_with_connection(move |conn| {
+            SqliteStore::insert_account(conn, &account, account_seed)
+        })
+        .await
+    }
+
+    async fn update_account(&self, account: &Account) -> Result<(), StoreError> {
+        let account = account.clone();
+
+        self.interact_with_connection(move |conn| SqliteStore::update_account(conn, &account))
+            .await
+    }
+
+    async fn get_account_ids(&self) -> Result<Vec<AccountId>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_account_ids).await
+    }
+
+    async fn get_account_headers(&self) -> Result<Vec<(AccountHeader, AccountStatus)>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_account_headers).await
+    }
+
+    async fn get_account_header(
+        &self,
+        account_id: AccountId,
+    ) -> Result<Option<(AccountHeader, AccountStatus)>, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::get_account_header(conn, account_id))
+            .await
+    }
+
+    async fn get_account_header_by_commitment(
+        &self,
+        account_commitment: Digest,
+    ) -> Result<Option<AccountHeader>, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_account_header_by_commitment(conn, account_commitment)
+        })
+        .await
+    }
+
+    async fn get_account(
+        &self,
+        account_id: AccountId,
+    ) -> Result<Option<AccountRecord>, StoreError> {
+        self.interact_with_connection(move |conn| SqliteStore::get_account(conn, account_id))
+            .await
+    }
+
+    async fn upsert_foreign_account_code(
+        &self,
+        account_id: AccountId,
+        code: AccountCode,
+    ) -> Result<(), StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::upsert_foreign_account_code(conn, account_id, &code)
+        })
+        .await
+    }
+
+    async fn get_foreign_account_code(
+        &self,
+        account_ids: Vec<AccountId>,
+    ) -> Result<BTreeMap<AccountId, AccountCode>, StoreError> {
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_foreign_account_code(conn, account_ids)
+        })
+        .await
+    }
+
+    async fn get_unspent_input_note_nullifiers(&self) -> Result<Vec<Nullifier>, StoreError> {
+        self.interact_with_connection(SqliteStore::get_unspent_input_note_nullifiers)
+            .await
+    }
 }
 
 // UTILS
@@ -65,4 +387,41 @@ pub fn u64_to_value(v: u64) -> Value {
         reason = "We store u64 as i64 as sqlite only allows the latter."
     )]
     Value::Integer(v as i64)
+}
+
+// TESTS
+// ================================================================================================
+
+#[cfg(test)]
+pub mod tests {
+    use std::boxed::Box;
+
+    use super::SqliteStore;
+    use crate::{store::Store, tests::create_test_store_path};
+
+    fn assert_send_sync<T: Send + Sync>() {}
+
+    #[test]
+    fn is_send_sync() {
+        assert_send_sync::<SqliteStore>();
+        assert_send_sync::<Box<dyn Store>>();
+    }
+
+    // Function that returns a `Send` future from a dynamic trait that must be `Sync`.
+    async fn dyn_trait_send_fut(store: Box<dyn Store>) {
+        // This wouldn't compile if `get_tracked_block_headers` doesn't return a `Send` future.
+        let res = store.get_tracked_block_headers().await;
+        assert!(res.is_ok());
+    }
+
+    #[tokio::test]
+    async fn future_is_send() {
+        let client = SqliteStore::new(create_test_store_path()).await.unwrap();
+        let client: Box<SqliteStore> = client.into();
+        tokio::task::spawn(async move { dyn_trait_send_fut(client).await });
+    }
+
+    pub(crate) async fn create_test_store() -> SqliteStore {
+        SqliteStore::new(create_test_store_path()).await.unwrap()
+    }
 }
