--- crates/rust-client/src/store/web_store/README.md
@@ -1,2 +1,0 @@
-## Formatting for .js files
-Please install the VSCode prettier extension and set as default formatter for js
