--- crates/rust-client/src/store/web_store/export/js_bindings.rs
@@ -1,8 +1,0 @@
-use wasm_bindgen::prelude::*;
-use wasm_bindgen_futures::{js_sys, wasm_bindgen};
-
-#[wasm_bindgen(module = "/src/store/web_store/js/export.js")]
-extern "C" {
-    #[wasm_bindgen(js_name = exportStore)]
-    pub fn idxdb_export_store() -> js_sys::Promise;
-}
