--- crates/rust-client/src/store/web_store/note/models.rs
@@ -1,43 +1,0 @@
-use alloc::{string::String, vec::Vec};
-
-use base64::{Engine as _, engine::general_purpose};
-use serde::{Deserialize, Deserializer, Serialize, de::Error};
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct InputNoteIdxdbObject {
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub assets: Vec<u8>,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub serial_number: Vec<u8>,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub inputs: Vec<u8>,
-    pub created_at: String,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub serialized_note_script: Vec<u8>,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub state: Vec<u8>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct OutputNoteIdxdbObject {
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub assets: Vec<u8>,
-    pub recipient_digest: String,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub metadata: Vec<u8>,
-    pub expected_height: u32,
-    #[serde(deserialize_with = "base64_to_vec_u8_required", default)]
-    pub state: Vec<u8>,
-}
-
-fn base64_to_vec_u8_required<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let base64_str: String = Deserialize::deserialize(deserializer)?;
-    general_purpose::STANDARD
-        .decode(&base64_str)
-        .map_err(|e| Error::custom(format!("Base64 decode error: {e}")))
-}
