+++ crates/web-client/Cargo.toml
@@ -22,6 +22,7 @@ workspace = true
 testing = ["miden-client/testing"]
 
 [dependencies]
+hex = { version = "0.4" }
 miden-client = { version = "0.10", path = "../rust-client", default-features = false, features = ["web", "web-tonic"] }
 miden-lib = { workspace = true }
 miden-objects = { workspace = true }
