+++ crates/web-client/src/export.rs
@@ -72,17 +72,18 @@ impl WebClient {
         }
     }
 
-    /// Retrieves the entire underlying web store and returns it as a JsValue
-    ///
-    /// Meant to be used in conjunction with the force_import_store method
-    #[wasm_bindgen(js_name = "exportStore")]
-    pub async fn export_store(&mut self) -> Result<JsValue, JsValue> {
-        let store = self.store.as_ref().ok_or(JsValue::from_str("Store not initialized"))?;
-        let export = store
-            .export_store()
-            .await
-            .map_err(|err| js_error_with_context(err, "failed to export store"))?;
+    // TODO(Maks) export OPFS
+    // Retrieves the entire underlying web store and returns it as a JsValue
+    //
+    // Meant to be used in conjunction with the force_import_store method
+    // #[wasm_bindgen(js_name = "exportStore")]
+    // pub async fn export_store(&mut self) -> Result<JsValue, JsValue> {
+    //     let store = self.store.as_ref().ok_or(JsValue::from_str("Store not initialized"))?;
+    //     let export = store
+    //         .export_store()
+    //         .await
+    //         .map_err(|err| js_error_with_context(err, "failed to export store"))?;
 
-        Ok(export)
-    }
+    //     Ok(export)
+    // }
 }
